+++ src/script/bridge/qscriptfunction.cpp
@@ -42,8 +42,8 @@ ASSERT_CLASS_FITS_IN_CELL(QT_PREPEND_NAMESPACE(QScript::FunctionWithArgWrapper))
 
 namespace QScript {
 
-const JSC::ClassInfo FunctionWrapper::info = { "QtNativeFunctionWrapper", &PrototypeFunction::info, nullptr, nullptr };
-const JSC::ClassInfo FunctionWithArgWrapper::info = { "QtNativeFunctionWithArgWrapper", &PrototypeFunction::info, nullptr, nullptr };
+const JSC::ClassInfo FunctionWrapper::info = { "QtNativeFunctionWrapper", &JSC::JSFunction::s_info, nullptr, nullptr };
+const JSC::ClassInfo FunctionWithArgWrapper::info = { "QtNativeFunctionWithArgWrapper", &JSC::JSFunction::s_info, nullptr, nullptr };
 
 FunctionWrapper::FunctionWrapper(JSC::ExecState *exec, int length, const JSC::Identifier &name,
    QScriptEngine::FunctionSignature function)
@@ -60,18 +60,17 @@ FunctionWrapper::~FunctionWrapper()
 JSC::ConstructType FunctionWrapper::getConstructData(JSC::ConstructData &consData)
 {
    consData.native.function = proxyConstruct;
-   consData.native.function.doNotCallDebuggerFunctionExit();
    return JSC::ConstructTypeHost;
 }
 
-JSC::JSValue FunctionWrapper::proxyCall(JSC::ExecState *exec, JSC::JSObject *callee,
-   JSC::JSValue thisObject, const JSC::ArgList &args)
+JSC::EncodedJSValue FunctionWrapper::proxyCall(JSC::ExecState *exec)
 {
-   FunctionWrapper *self = static_cast<FunctionWrapper *>(callee);
+   FunctionWrapper *self = static_cast<FunctionWrapper *>(exec->callee());
    QScriptEnginePrivate *eng_p = QScript::scriptEngineFromExec(exec);
+   JSC::ArgList args(exec);
 
    JSC::ExecState *oldFrame = eng_p->currentFrame;
-   eng_p->pushContext(exec, thisObject, args, callee);
+   eng_p->pushContext(exec, exec->hostThisValue(), args, self);
    QScriptContext *ctx = eng_p->contextForFrame(eng_p->currentFrame);
 
    QScriptValue result = self->data->function(ctx, QScriptEnginePrivate::get(eng_p));
@@ -82,24 +81,20 @@ JSC::JSValue FunctionWrapper::proxyCall(JSC::ExecState *exec, JSC::JSObject *cal
    eng_p->popContext();
    eng_p->currentFrame = oldFrame;
 
-   return eng_p->scriptValueToJSCValue(result);
+   return JSC::JSValue::encode(eng_p->scriptValueToJSCValue(result));
 }
 
-JSC::JSObject *FunctionWrapper::proxyConstruct(JSC::ExecState *exec, JSC::JSObject *callee,
-   const JSC::ArgList &args)
+JSC::EncodedJSValue FunctionWrapper::proxyConstruct(JSC::ExecState *exec)
 {
-   FunctionWrapper *self = static_cast<FunctionWrapper *>(callee);
+   FunctionWrapper *self = static_cast<FunctionWrapper *>(exec->callee());
    QScriptEnginePrivate *eng_p = QScript::scriptEngineFromExec(exec);
+   JSC::ArgList args(exec);
 
    JSC::ExecState *oldFrame = eng_p->currentFrame;
-   eng_p->pushContext(exec, JSC::JSValue(), args, callee, true);
+   eng_p->pushContext(exec, JSC::JSValue(), args, self, true);
    QScriptContext *ctx = eng_p->contextForFrame(eng_p->currentFrame);
 
    QScriptValue result = self->data->function(ctx, QScriptEnginePrivate::get(eng_p));
-
-   if (JSC::Debugger *debugger = eng_p->originalGlobalObject()->debugger()) {
-      debugger->functionExit(QScriptValuePrivate::get(result)->jscValue, -1);
-   }
 
    if (!result.isObject()) {
       result = ctx->thisObject();
@@ -108,7 +103,7 @@ JSC::JSObject *FunctionWrapper::proxyConstruct(JSC::ExecState *exec, JSC::JSObje
    eng_p->popContext();
    eng_p->currentFrame = oldFrame;
 
-   return JSC::asObject(eng_p->scriptValueToJSCValue(result));
+   return JSC::JSValue::encode(eng_p->scriptValueToJSCValue(result));
 }
 
 FunctionWithArgWrapper::FunctionWithArgWrapper(JSC::ExecState *exec, int length, const JSC::Identifier &name,
@@ -131,14 +126,14 @@ JSC::ConstructType FunctionWithArgWrapper::getConstructData(JSC::ConstructData &
    return JSC::ConstructTypeHost;
 }
 
-JSC::JSValue FunctionWithArgWrapper::proxyCall(JSC::ExecState *exec, JSC::JSObject *callee,
-   JSC::JSValue thisObject, const JSC::ArgList &args)
+JSC::EncodedJSValue FunctionWithArgWrapper::proxyCall(JSC::ExecState *exec)
 {
-   FunctionWithArgWrapper *self = static_cast<FunctionWithArgWrapper *>(callee);
+   FunctionWithArgWrapper *self = static_cast<FunctionWithArgWrapper *>(exec->callee());
    QScriptEnginePrivate *eng_p = QScript::scriptEngineFromExec(exec);
+   JSC::ArgList args(exec);
 
    JSC::ExecState *oldFrame = eng_p->currentFrame;
-   eng_p->pushContext(exec, thisObject, args, callee);
+   eng_p->pushContext(exec, exec->hostThisValue(), args, self);
    QScriptContext *ctx = eng_p->contextForFrame(eng_p->currentFrame);
 
    QScriptValue result = self->data->function(ctx, QScriptEnginePrivate::get(eng_p), self->data->arg);
@@ -146,17 +141,17 @@ JSC::JSValue FunctionWithArgWrapper::proxyCall(JSC::ExecState *exec, JSC::JSObje
    eng_p->popContext();
    eng_p->currentFrame = oldFrame;
 
-   return eng_p->scriptValueToJSCValue(result);
+   return JSC::JSValue::encode(eng_p->scriptValueToJSCValue(result));
 }
 
-JSC::JSObject *FunctionWithArgWrapper::proxyConstruct(JSC::ExecState *exec, JSC::JSObject *callee,
-   const JSC::ArgList &args)
+JSC::EncodedJSValue FunctionWithArgWrapper::proxyConstruct(JSC::ExecState *exec)
 {
-   FunctionWithArgWrapper *self = static_cast<FunctionWithArgWrapper *>(callee);
+   FunctionWithArgWrapper *self = static_cast<FunctionWithArgWrapper *>(exec->callee());
    QScriptEnginePrivate *eng_p = QScript::scriptEngineFromExec(exec);
+   JSC::ArgList args(exec);
 
    JSC::ExecState *oldFrame = eng_p->currentFrame;
-   eng_p->pushContext(exec, JSC::JSValue(), args, callee, true);
+   eng_p->pushContext(exec, JSC::JSValue(), args, self, true);
    QScriptContext *ctx = eng_p->contextForFrame(eng_p->currentFrame);
 
    QScriptValue result = self->data->function(ctx, QScriptEnginePrivate::get(eng_p), self->data->arg);
@@ -167,7 +162,7 @@ JSC::JSObject *FunctionWithArgWrapper::proxyConstruct(JSC::ExecState *exec, JSC:
    eng_p->popContext();
    eng_p->currentFrame = oldFrame;
 
-   return JSC::asObject(eng_p->scriptValueToJSCValue(result));
+   return JSC::JSValue::encode(eng_p->scriptValueToJSCValue(result));
 }
 
 } // namespace